requirements.txt
PySide6
onnxruntime
tokenizers
numpy
libsql
PyPDF2
pyobjc-framework-AVFoundation
pyobjc-framework-ScreenCaptureKit
pyobjc-framework-Speech
